requirements.txt
pySerial~=3.5
xmodem~=0.4.6